Se refactorizaron los metodos referentes a filtros
