+++ lib/menu_pendientes.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:ing_software_grupo4/handlers/report_handler.dart';
 import 'package:ing_software_grupo4/modelos/filter.dart';
+import 'package:ing_software_grupo4/modelos/filter_utils.dart';
 import 'package:ing_software_grupo4/modelos/modo.dart';
-import 'package:ing_software_grupo4/modelos/tipo_reporte.dart';
 import 'package:ing_software_grupo4/tarjeta_reporte.dart';
 
 class MenuPendientes extends StatefulWidget {
@@ -78,133 +78,11 @@ class _MenuPendientesState extends State<MenuPendientes> {
   }
 
   Future<void> _openFilterDialog() async {
-    // Definir todas las categorías disponibles siempre
-    final Set<String> availableTags = {
-      'Celular',
-      'Notebook / Laptop',
-      'Tablet',
-      'Audífonos',
-      'Cargador / Cable',
-      'Reloj inteligente',
-      'Lentes',
-      'Llaves',
-      'Billetera',
-      'Cartera',
-      'Paraguas',
-      'Botella de agua',
-      'Libro',
-      'Mochila',
-      'Chaqueta',
-      'Gorro',
-      'Bufanda',
-      'Guantes',
-      'Calculadora',
-      'USB / Pendrive',
-      'Tarjeta de memoria',
-      'Mouse',
-      'Teclado',
-      'Otro',
-    };
-
-    final Set<String> availableColors = {};
-    try {
-      availableColors.addAll(colorNameToHex.keys);
-    } catch (_) {}
-
-    final Set<String> availableTipos = {};
-    try {
-      availableTipos.addAll(
-          TipoReporte.values.where((t) => t.name != 'administracion').map((t) => t.name));
-    } catch (_) {}
-
-    final tempTags = Set<String>.from(_activeTagFilters);
-    final tempColors = Set<String>.from(_activeColorFilters);
-    final tempTipos = Set<String>.from(_activeTipoFilters);
-
-    final result = await showDialog<Map<String, Set<String>>>(
-      context: context,
-      builder: (context) {
-        return AlertDialog(
-          title: const Text('Filtros'),
-          content: SingleChildScrollView(
-            child: SizedBox(
-              width: double.maxFinite,
-              child: StatefulBuilder(
-                builder: (context, setState) {
-                  return Column(
-                    mainAxisSize: MainAxisSize.min,
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      const Text('Categorías',
-                          style: TextStyle(fontWeight: FontWeight.w700)),
-                      const SizedBox(height: 6),
-                      ...availableTags.map((name) {
-                        return CheckboxListTile(
-                          title: Text(name),
-                          value: tempTags.contains(name),
-                          onChanged: (v) => setState(() {
-                            if (v == true) {
-                              tempTags.add(name);
-                            } else {
-                              tempTags.remove(name);
-                            }
-                          }),
-                        );
-                      }),
-                      const Divider(),
-                      const Text('Colores',
-                          style: TextStyle(fontWeight: FontWeight.w700)),
-                      const SizedBox(height: 6),
-                      ...availableColors.map((name) {
-                        return CheckboxListTile(
-                          title: Text(prettifyColorName(name)),
-                          value: tempColors.contains(name),
-                          onChanged: (v) => setState(() {
-                            if (v == true) {
-                              tempColors.add(name);
-                            } else {
-                              tempColors.remove(name);
-                            }
-                          }),
-                        );
-                      }),
-                      const Divider(),
-                      const Text('Tipos',
-                          style: TextStyle(fontWeight: FontWeight.w700)),
-                      const SizedBox(height: 6),
-                      ...availableTipos.map((name) {
-                        return CheckboxListTile(
-                          title: Text(name[0].toUpperCase() + name.substring(1)),
-                          value: tempTipos.contains(name),
-                          onChanged: (v) => setState(() {
-                            if (v == true) {
-                              tempTipos.add(name);
-                            } else {
-                              tempTipos.remove(name);
-                            }
-                          }),
-                        );
-                      }),
-                    ],
-                  );
-                },
-              ),
-            ),
-          ),
-          actions: [
-            TextButton(
-                onPressed: () => Navigator.pop(context),
-                child: const Text('Cancelar')),
-            TextButton(
-                onPressed: () => Navigator.pop(context, {
-                      'tags': tempTags,
-                      'colors': tempColors,
-                      'tipos': tempTipos,
-                    }),
-                child: const Text('Aplicar')),
-          ],
-        );
-      },
+    final result = await openFilterDialog(
+      context,
+      activeTagFilters: _activeTagFilters,
+      activeColorFilters: _activeColorFilters,
+      activeTipoFilters: _activeTipoFilters,
     );
 
     if (result != null) {
@@ -220,27 +98,6 @@ class _MenuPendientesState extends State<MenuPendientes> {
           ..addAll(result['tipos'] ?? {});
       });
     }
-  }
-
-  String prettifyColorName(String colorName) {
-    return colorName[0].toUpperCase() +
-        colorName.substring(1).replaceAll('_', ' ');
-  }
-
-  Map<String, int> get colorNameToHex {
-    return {
-      'rojo': 0xFFFF0000,
-      'azul': 0xFF0000FF,
-      'verde': 0xFF00FF00,
-      'amarillo': 0xFFFFFF00,
-      'negro': 0xFF000000,
-      'blanco': 0xFFFFFFFF,
-      'gris': 0xFF808080,
-      'naranja': 0xFFFFA500,
-      'morado': 0xFF800080,
-      'rosa': 0xFFFFC0CB,
-      'cafe': 0xFF8B4513,
-    };
   }
 }
