+++ lib/menu_reportes.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:ing_software_grupo4/modelos/filter.dart';
+import 'package:ing_software_grupo4/modelos/filter_utils.dart';
 import 'package:ing_software_grupo4/modelos/reporte.dart';
 import 'package:ing_software_grupo4/modelos/tag.dart';
-import 'package:ing_software_grupo4/modelos/tipo_reporte.dart';
 import 'package:ing_software_grupo4/tarjeta_reporte.dart';
 import 'package:ing_software_grupo4/handlers/report_handler.dart';
 import 'package:ing_software_grupo4/handlers/session_handler.dart';
@@ -84,146 +84,11 @@ class _MenuReportesState extends State<MenuReportes> {
   }
 
   Future<void> _openFilterDialog() async {
-    // Definir todas las categorías disponibles siempre
-    final Set<String> availableTags = {
-      'Celular',
-      'Notebook / Laptop',
-      'Tablet',
-      'Audífonos',
-      'Cargador / Cable',
-      'Reloj inteligente',
-      'Lentes',
-      'Llaves',
-      'Billetera',
-      'Cartera',
-      'Paraguas',
-      'Botella de agua',
-      'Libro',
-      'Mochila',
-      'Chaqueta',
-      'Gorro',
-      'Bufanda',
-      'Guantes',
-      'Calculadora',
-      'USB / Pendrive',
-      'Tarjeta de memoria',
-      'Mouse',
-      'Teclado',
-      'Otro',
-    };
-
-    final Set<String> availableColors = {};
-    try {
-      availableColors.addAll(colorNameToHex.keys);
-    } catch (_) {}
-
-    final Set<String> availableTipos = {};
-    try {
-      availableTipos.addAll(
-        TipoReporte.values
-            .where((t) => t.name != 'administracion')
-            .map((t) => t.name),
-      );
-    } catch (_) {}
-
-    final tempTags = Set<String>.from(_activeTagFilters);
-    final tempColors = Set<String>.from(_activeColorFilters);
-    final tempTipos = Set<String>.from(_activeTipoFilters);
-
-    final result = await showDialog<Map<String, Set<String>>>(
-      context: context,
-      builder: (context) {
-        return AlertDialog(
-          title: const Text('Filtros'),
-          content: SingleChildScrollView(
-            child: SizedBox(
-              width: double.maxFinite,
-              child: StatefulBuilder(
-                builder: (context, setState) {
-                  return Column(
-                    mainAxisSize: MainAxisSize.min,
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      const Text(
-                        'Categorí­as',
-                        style: TextStyle(fontWeight: FontWeight.w700),
-                      ),
-                      const SizedBox(height: 6),
-                      ...availableTags.map((name) {
-                        return CheckboxListTile(
-                          title: Text(name),
-                          value: tempTags.contains(name),
-                          onChanged: (v) => setState(() {
-                            if (v == true) {
-                              tempTags.add(name);
-                            } else {
-                              tempTags.remove(name);
-                            }
-                          }),
-                        );
-                      }),
-                      const Divider(),
-                      const Text(
-                        'Colores',
-                        style: TextStyle(fontWeight: FontWeight.w700),
-                      ),
-                      const SizedBox(height: 6),
-                      ...availableColors.map((name) {
-                        return CheckboxListTile(
-                          title: Text(prettifyColorName(name)),
-                          value: tempColors.contains(name),
-                          onChanged: (v) => setState(() {
-                            if (v == true) {
-                              tempColors.add(name);
-                            } else {
-                              tempColors.remove(name);
-                            }
-                          }),
-                        );
-                      }),
-                      const Divider(),
-                      const Text(
-                        'Tipos',
-                        style: TextStyle(fontWeight: FontWeight.w700),
-                      ),
-                      const SizedBox(height: 6),
-                      ...availableTipos.map((name) {
-                        return CheckboxListTile(
-                          title: Text(
-                            name[0].toUpperCase() + name.substring(1),
-                          ),
-                          value: tempTipos.contains(name),
-                          onChanged: (v) => setState(() {
-                            if (v == true) {
-                              tempTipos.add(name);
-                            } else {
-                              tempTipos.remove(name);
-                            }
-                          }),
-                        );
-                      }),
-                    ],
-                  );
-                },
-              ),
-            ),
-          ),
-          actions: [
-            TextButton(
-              onPressed: () => Navigator.pop(context),
-              child: const Text('Cancelar'),
-            ),
-            TextButton(
-              onPressed: () => Navigator.pop(context, {
-                'tags': tempTags,
-                'colors': tempColors,
-                'tipos': tempTipos,
-              }),
-              child: const Text('Aplicar'),
-            ),
-          ],
-        );
-      },
+    final result = await openFilterDialog(
+      context,
+      activeTagFilters: _activeTagFilters,
+      activeColorFilters: _activeColorFilters,
+      activeTipoFilters: _activeTipoFilters,
     );
 
     if (result != null) {
@@ -239,13 +104,6 @@ class _MenuReportesState extends State<MenuReportes> {
           ..addAll(result['tipos'] ?? {});
       });
     }
-  }
-
-  String prettifyColorName(String key) {
-    return key
-        .split('_')
-        .map((s) => s.isEmpty ? s : (s[0].toUpperCase() + s.substring(1)))
-        .join(' ');
   }
 }
 
@@ -410,36 +268,3 @@ class ListaReportes extends StatelessWidget {
     );
   }
 }
-
-const Map<String, String> colorNameToHex = {
-  'rojo': '#FF0000',
-  'verde': '#00FF00',
-  'azul': '#0000FF',
-  'amarillo': '#FFFF00',
-  'naranja': '#FFA500',
-  'morado': '#800080',
-  'rosa': '#FFC0CB',
-  'negro': '#000000',
-  'blanco': '#FFFFFF',
-  'gris': '#808080',
-  'cafe': '#8B4513',
-  'celeste': '#87CEEB',
-  'turquesa': '#40E0D0',
-  'plateado': '#C0C0C0',
-  'dorado': '#FFD700',
-  'beige': '#F5F5DC',
-  'marron': '#A52A2A',
-  'verde_oscuro': '#006400',
-  'azul_marino': '#000080',
-  'violeta': '#8A2BE2',
-  'fucsia': '#FF00FF',
-  'crema': '#FFFDD0',
-  'coral': '#FF7F50',
-  'salmon': '#FA8072',
-  'lavanda': '#E6E6FA',
-  'menta': '#98FF98',
-  'durazno': '#FFE5B4',
-  'bordo': '#B80F0A',
-  'burdeos': '#800020',
-  'Otro': '#FFFFFF',
-};
